+++ packages/@webex/plugin-meetings/test/unit/spec/stats-analyzer/index.js
@@ -7,6 +7,7 @@ import {ConnectionState} from '@webex/internal-media-core';
 import {StatsAnalyzer, EVENTS} from '../../../../src/statsAnalyzer';
 import NetworkQualityMonitor from '../../../../src/networkQualityMonitor';
 import testUtils from '../../../utils/testUtils';
+import {MEDIA_DEVICES, _UNKNOWN_} from '@webex/plugin-meetings/src/constants';
 
 const {assert} = chai;
 
@@ -119,6 +120,7 @@ describe('plugin-meetings', () => {
           audio: {
             senders: [
               {
+                localTrackLabel: 'fake-microphone',
                 report: [
                   {
                     type: 'outbound-rtp',
@@ -173,10 +175,11 @@ describe('plugin-meetings', () => {
           video: {
             senders: [
               {
+                localTrackLabel: 'fake-camera',
                 report: [
                   {
                     type: 'outbound-rtp',
-                    framesSent: 0,
+                    framesSent: 1500,
                     bytesSent: 1,
                   },
                   {
@@ -206,7 +209,7 @@ describe('plugin-meetings', () => {
                     bytesReceived: 1,
                     frameHeight: 720,
                     frameWidth: 1280,
-                    framesReceived: 1,
+                    framesReceived: 1500,
                   },
                   {
                     type: 'candidate-pair',
@@ -412,6 +415,37 @@ describe('plugin-meetings', () => {
 
         assert.strictEqual(mqeData.audioTransmit[0].common.transportType, 'TLS');
         assert.strictEqual(mqeData.videoReceive[0].common.transportType, 'TLS');
+      });
+
+      it('emits the correct peripherals in MEDIA_QUALITY events', async () => {
+        await startStatsAnalyzer({expected: {receiveVideo: true}});
+
+        await progressTime();
+
+        assert.strictEqual(mqeData.intervalMetadata.peripherals.find((val) => val.name === MEDIA_DEVICES.MICROPHONE).information, 'fake-microphone');
+        assert.strictEqual(mqeData.intervalMetadata.peripherals.find((val) => val.name === MEDIA_DEVICES.CAMERA).information, 'fake-camera');
+      });
+
+      it('emits the correct peripherals in MEDIA_QUALITY events when localTrackLabel is undefined', async () => {
+        fakeStats.audio.senders[0].localTrackLabel = undefined;
+        fakeStats.video.senders[0].localTrackLabel = undefined;
+
+        await startStatsAnalyzer({expected: {receiveVideo: true}});
+
+        await progressTime();
+
+        assert.strictEqual(mqeData.intervalMetadata.peripherals.find((val) => val.name === MEDIA_DEVICES.MICROPHONE).information, _UNKNOWN_);
+        assert.strictEqual(mqeData.intervalMetadata.peripherals.find((val) => val.name === MEDIA_DEVICES.CAMERA).information, _UNKNOWN_);
+      });
+
+      it('emits the correct frameRate', async () => {
+        await startStatsAnalyzer({expected: {receiveVideo: true}});
+
+        await progressTime();
+        assert.strictEqual(mqeData.videoReceive[0].streams[0].common.receivedFrameRate, 25);
+        fakeStats.video.receivers[0].framesReceived = 3000;
+        await progressTime();
+        assert.strictEqual(mqeData.videoReceive[0].streams[0].common.receivedFrameRate, 25);
       });
     });
   });